Airports filtered by Minimum Number of Departures
Each Aircraft should have a minimum amount of departures at the airport to deliver enough datapoints for a sufficient analysis. 
Coordinates of the start of each climb (tilmestep = 0) were merged by rounding to one digit. This follows the assumption, that two major international airports do not exist within a distance of 0,1°.